+++ public/CHANGELOG.md
@@ -4,3 +4,4 @@
 - 添加文档面包屑 | Add document breadcrumb.
 - 添加只读/编辑切换开关 | Add read-only/edit switch.
 - 添加文档重命名功能 | Add document rename function.
+- 添加原生属性保存功能 | Add native attribute save function.